all cost working and displaying the same value on checkout

+++ app/redux/authentication.js
@@ -124,7 +124,7 @@ export const initialLogInAttempt = () => {
           );
           // add the items to the redux store
           let price = localStorageItems.reduce((acc, item) => {
-            acc += parseInt(item.unitPrice);
+            acc += parseFloat(item.unitPrice);
             return acc;
           }, 0);
           return dispatch(setGuestItemsToCart(localStorageItems, price));
