+++ app/redux/cart.js
@@ -184,7 +184,7 @@ const cartReducer = (state = initialState, action) => {
         // I needed to uncomment the below out in order to make the cart -JH
         // update when a user is logged out
         items: [...state.items, action.orderItem],
-        orderTotal: (state.orderTotal += parseInt(action.orderItem.unitPrice))
+        orderTotal: (state.orderTotal += parseFloat(action.orderItem.unitPrice))
       };
 
     // DONT NEED THIS ANYMORE EITHER SINCE WE WONT BE KEEP QUANTITY. BUT LEAVING IT HERE JUST INCASE.
@@ -205,7 +205,7 @@ const cartReducer = (state = initialState, action) => {
         items: state.items.filter(
           orderItem => orderItem.id !== action.orderItem.id
         ),
-        orderTotal: (state.orderTotal -= parseInt(action.orderItem.unitPrice))
+        orderTotal: (state.orderTotal -= parseFloat(action.orderItem.unitPrice))
       };
     case EMPTY_CART:
       return {
